+++ Sources/TuistGenerator/Linter/SettingsLinter.swift
@@ -22,6 +22,10 @@ final class SettingsLinter: SettingsLinting {
         if let settings = target.settings {
             issues.append(contentsOf: lintConfigFilesExist(settings: settings))
         }
+
+        if let deploymentTarget = target.deploymentTarget {
+            issues.append(contentsOf: lint(platform: target.platform, isCompatibleWith: deploymentTarget))
+        }
         return issues
     }
 
@@ -46,6 +50,17 @@ final class SettingsLinter: SettingsLinting {
     private func lintNonEmptyConfig(project: Project) -> [LintingIssue] {
         guard !project.settings.configurations.isEmpty else {
             return [LintingIssue(reason: "The project at path \(project.path.pathString) has no configurations", severity: .error)]
+        }
+        return []
+    }
+
+    // TODO_MAJOR_CHANGE: Merge deploymentTarget and platform arguments together.
+    private func lint(platform: Platform, isCompatibleWith deploymentTarget: DeploymentTarget) -> [LintingIssue] {
+        let issue = LintingIssue(reason: "Found an inconsistency between a platform `\(platform.caseValue)` and deployment target `\(deploymentTarget.platform)`", severity: .error)
+
+        switch deploymentTarget {
+        case .iOS: if platform != .iOS { return [issue] }
+        case .macOS: if platform != .macOS { return [issue] }
         }
         return []
     }