+++ Tests/ProjectDescriptionTests/TargetTests.swift
@@ -11,6 +11,7 @@ final class TargetTests: XCTestCase {
                              product: .app,
                              productName: "product_name",
                              bundleId: "bundle_id",
+                             deploymentTarget: .iOS(targetVersion: "13.1", devices: [.iphone, .ipad]),
                              infoPlist: "info.plist",
                              sources: "sources/*",
                              resources: "resources/*",
@@ -37,6 +38,11 @@ final class TargetTests: XCTestCase {
 
         let expected = """
         {
+            "deploymentTarget": {
+               "kind": "iOS",
+               "version": "13.1",
+               "deploymentDevices": 3
+            },
             "headers": {
                 "public": { "globs": ["public\\/*"] },
                 "private": { "globs": ["private\\/*"] },